% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IV_MH_Interval_Censored.R
\name{IV_MH_IC}
\alias{IV_MH_IC}
\title{Parametric Bayesian Instrumental Variable (PBIV) Methods
IV analysis function with bivariate normal random errors
with multiple instruments and multiple observed potential confounders.
This function deals with the arbitrary censoring version of Li-Lu's PBIV method published
in \emph{Stats. in Med.} \url{https://www.ncbi.nlm.nih.gov/pmc/articles/PMC4314427/}.}
\usage{
IV_MH_IC(L, R, d, X, G, U = NULL, m, wid, init = NULL, prior_1, prior_2)
}
\arguments{
\item{L}{left-observed time to event or censoring}

\item{R}{right-observed time to event or censoring}

\item{d}{censoring status, 4=event, 3=right-censored, 2=interval-censored, 1=left-censored}

\item{X}{covariate of interest}

\item{G}{a matrix of n x kG if multiple instruments are used;}

\item{U}{a matrix of n x kU if multiple observed confounders are used; the default is NULL
where n is sample size, kG is number of instruments, kU is number of observed confounders
A total of (6+kG+2*kU) parameters to estimate: a0, a1 (length=kG), a2 (length=kU), sig1, b0, b1, b2 (length=kU), sig2, and rho}

\item{m}{number of iterations}

\item{wid}{a vector of the random walk width for (a0,a1,a2,sig1,b0,b1,b2,sig2,rho)}

\item{init}{a vector of the initial values for (a0,a1,a2,sig1,b0,b1,b2,sig2,rho), default value is NULL}

\item{prior_1}{a vector of the first parameter of the priors for (a0,a1,a2,sig1,b0,b1,b2,sig2):
mean of the normal priors for a0,a1,a2,b0,b1,b2; shape parameter of the inverse-gamma priors for sig1, sig2}

\item{prior_2}{a vector of the second parameter of the priors for (a0,a1,a2,sig1,b0,b1,b2,sig2):
variance of the normal priors for a0,a1,a2,b0,b1,b2; scale parameter of the inverse-gamma priors for sig1, sig2}
}
\value{
A list containing traces and accept rates
}
\description{
Parametric Bayesian Instrumental Variable (PBIV) Methods
IV analysis function with bivariate normal random errors
with multiple instruments and multiple observed potential confounders.
This function deals with the arbitrary censoring version of Li-Lu's PBIV method published
in \emph{Stats. in Med.} \url{https://www.ncbi.nlm.nih.gov/pmc/articles/PMC4314427/}.
}
\author{
Xuyang Lu \href{mailto:xylu@ucla.edu}{xylu@ucla.edu}; Elvis Cui \href{mailto:elviscuihan@g.ucla.edu}{elviscuihan@g.ucla.edu}
Updated on 2022-07-22
}
